Move grant middleware in controller action
