--- packages/strapi-plugin-users-permissions/middlewares/provider/defaults.json
@@ -1,5 +1,0 @@
-{
-  "provider": {
-    "enabled": true
-  }
-}
